pyproject.toml
[project]
name = "digifactory-scraper"
version = "0.1.0"
description = "Scraper pour DigiFactory avec stockage Supabase"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.25.0",
    "selectolax>=0.3.0",
    "orjson>=3.9.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.2.0",
    "aiosqlite>=0.19.0",
]

[project.scripts]
digifactory-scraper = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

